requirements.txt
python-dotenv~=0.20.0
disnake~=2.4.0
requests~=2.27.1
selenium~=4.1.3
beautifulsoup4~=4.11.1
seleniumbase~=2.5.2